+++ pyquil/api/_abstract_compiler.py
@@ -28,6 +28,7 @@ from pyquil.quantum_processor import AbstractQuantumProcessor
 from pyquil.quil import Program
 from pyquil.quilatom import ExpressionDesignator, MemoryReference
 from pyquil.quilbase import Gate
+from pyquil._version import DOCS_URL
 from qcs_api_client.client import QCSClientConfiguration
 from rpcq.messages import NativeQuilMetadata, ParameterAref, ParameterSpec
 
@@ -149,7 +150,8 @@ class AbstractCompiler(ABC):
             raise QuilcNotRunning(
                 f"Request to quilc at {self._compiler_client.base_url} timed out. "
                 "This could mean that quilc is not running, is not reachable, or is "
-                "responding slowly."
+                "responding slowly. See the Troubleshooting Guide: "
+                "{DOCS_URL}/troubleshooting.html"
             )
 
     @abstractmethod